[build] add back -Werror

+++ src/mbgl/geometry/collision_box_buffer.cpp
@@ -6,8 +6,8 @@
 
 namespace mbgl {
 
-size_t CollisionBoxVertexBuffer::add(int16_t x, int16_t y, float ox, float oy, float maxzoom, float placementZoom) {
-    const size_t idx = index();
+uint32_t CollisionBoxVertexBuffer::add(int16_t x, int16_t y, float ox, float oy, float maxzoom, float placementZoom) {
+    const uint32_t idx = index();
     void *data = addElement();
 
     int16_t *shorts = static_cast<int16_t *>(data);
